update form hidden field value
