+++ pages/former-ramona.njk
@@ -64,7 +64,7 @@ locale: en
       <p>Your comment can cover as many topics as you want. There is no character limit. Please submit only one comment.</p>
 
       <form method="post" class="stack" action="#" id="comment-form">
-        <input type="hidden" id="comment-form-name" "former-ramona" />
+        <input type="hidden" id="comment-form-name" value="former-ramona" />
         <label for="commentor-name">Name</label>
         <input type="text" id="commentor-name">
         <div class="form-error commentor-name-field">Enter your name</div>
